+++ packages/aws-lambda-rest/src/handler.ts
@@ -1,9 +1,10 @@
 import { attachRestMethods } from './methods'
 import { functions } from '@mondrian-framework/module'
 import { rest, utils } from '@mondrian-framework/rest'
+import { replaceLast } from '@mondrian-framework/utils'
 import { APIGatewayProxyHandlerV2 } from 'aws-lambda'
 import fs from 'fs'
-import lambdaApi, { Request, Response } from 'lambda-api'
+import lambdaApi, { Request, Response, API } from 'lambda-api'
 import path from 'path'
 import { getAbsoluteFSPath } from 'swagger-ui-dist'
 
@@ -13,28 +14,32 @@ export function build<const Fs extends functions.Functions, const ContextInput>(
   api,
   context,
   error,
+  customize,
   ...args
 }: {
   api: rest.Api<Fs, ContextInput>
   context: (serverContext: Context) => Promise<ContextInput>
   error?: rest.ErrorHandler<Fs, Context>
   options: Partial<rest.ServeOptions>
+  customize?: (server: API) => void
 }): APIGatewayProxyHandlerV2 {
   utils.assertApiValidity(api)
-  const server = lambdaApi()
+  const server = lambdaApi({ base: '' })
   const options = { ...rest.DEFAULT_SERVE_OPTIONS, ...args.options }
   if (options.introspection) {
-    const introspectionPath = options.introspection.path
+    const introspectionPath = options.introspection.path.endsWith('/')
+      ? options.introspection.path
+      : `${options.introspection.path}/`
     const indexContent = fs
       .readFileSync(path.join(getAbsoluteFSPath(), 'swagger-initializer.js'))
       .toString()
-      .replace('https://petstore.swagger.io/v2/swagger.json', `${introspectionPath}/v${api.version}/schema.json`)
-    server.get(`${introspectionPath}/swagger-initializer.js`, (req: Request, res: Response) => res.send(indexContent))
-    server.get(`${introspectionPath}`, (req: Request, res: Response) => {
-      res.redirect(`${introspectionPath}/index.html`)
+      .replace('https://petstore.swagger.io/v2/swagger.json', `${introspectionPath}v${api.version}/schema.json`)
+    server.get(`${introspectionPath}swagger-initializer.js`, (_: Request, res: Response) => res.send(indexContent))
+    server.get(`${introspectionPath}`, (_: Request, res: Response) => {
+      res.redirect(`${introspectionPath}index.html`)
     })
     const cache: Map<string, unknown> = new Map()
-    server.get(`${introspectionPath}/:v/schema.json`, (req: Request, res: Response) => {
+    server.get(`${introspectionPath}:v/schema.json`, (req: Request, res: Response) => {
       const v = (req.params as Record<string, string>).v
       const version = Number(v.replace('v', ''))
       if (Number.isNaN(version) || !Number.isInteger(version) || version < 1 || version > api.version) {
@@ -50,17 +55,20 @@ export function build<const Fs extends functions.Functions, const ContextInput>(
       return schema
     })
     // file deepcode ignore NoRateLimitingForExpensiveWebOperation: could disable this by disabling introspection in production environment
-    server.get(`${introspectionPath}/*`, (req: Request, res: Response) => {
+    server.get(`${introspectionPath}*`, (req: Request, res: Response) => {
       //avoid path traversal
       if (req.path.match(/\.\.\//g) !== null) {
         res.status(404)
         return
       }
       const file = `${getAbsoluteFSPath()}/${req.path}`
-      const path = file.replace(`${introspectionPath}/`, '')
+      const path = replaceLast(file, `${introspectionPath}`, '')
       res.sendFile(path)
     })
   }
   attachRestMethods({ api, context, server, error })
+  if (customize) {
+    customize(server)
+  }
   return (event, context) => server.run(event, context)
 }