+++ packages/model/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mondrian-framework/model
 
+## 2.0.15
+
+### Patch Changes
+
+- Minor improvements
+- Updated dependencies
+  - @mondrian-framework/utils@2.0.15
+
 ## 2.0.14
 
 ### Patch Changes